==> 0.FLUXNET_retrieval.ipynb <==
import pandas as pd
import os

### This file will concat all of the daily fluxnet files for all of the sites

### downloaded from: https://fluxnet.org/data/

directory='/Users/abigailbase/PROJECT FILES/RP Code Files'

#function to look through all of the downloaded files and extract only the ones with DD obseervations

def combine_files_by_timestamp(directory, final_output_directory):
    files_with_timestamp = []

    # all files in the output directory
    for dirpath, _, filenames in os.walk(directory):
        for filename in filenames:
            if filename.endswith('.csv'):
                file_path = os.path.join(dirpath, filename)
                
                # extractsite name from the filename
                site_name = filename.split('_')[1]
                
                try:
                    df = pd.read_csv(file_path)
                except Exception as e:
                    print(f"Error reading {file_path}: {e}")
                    continue
                
                # add SITE_ID column
                df['SITE_ID'] = site_name
                
                # check for timestamp column
                if 'TIMESTAMP' in df.columns:
                    files_with_timestamp.append(df)
    
    #  final output directory if it doesn't exist
    if not os.path.exists(final_output_directory):
        os.makedirs(final_output_directory)
        
    # combine files with timestamp
    if files_with_timestamp:
        combined_df_with_timestamp = pd.concat(files_with_timestamp, ignore_index=True)
        combined_with_timestamp_path = os.path.join(final_output_directory, 'combined_with_timestamp3.csv')
        combined_df_with_timestamp.to_csv(combined_with_timestamp_path, index=False)
        print(f"Combined files with timestamp saved to {combined_with_timestamp_path}")


final_output_directory = '/Users/abigailbase/PROJECT FILES/Final3'

combine_files_by_timestamp(directory, final_output_directory)

dtype = {
    75: 'str',  # or 'float', 'int', etc., based on your data
    76: 'str'   # or 'float', 'int', etc., based on your data
}


see=pd.read_csv('/Users/abigailbase/PROJECT FILES/Final3/combined_with_timestamp3.csv',dtype=dtype)

see['SITE_ID'].unique()

see=see.drop(columns=['TA_F_QC','SW_IN_POT','LW_IN_F_QC','SW_IN_F_QC','LW_IN_F_QC','VPD_F_QC'])

see.head()

see=see.drop(columns=['PA_F_QC','P_F_QC'])

print(see.columns)

var_list=df.columns.tolist()

see=see.drop(columns=['TS_F_MDS_1_QC','SWC_F_MDS_1_QC'])

columns_to_drop = [col for col in see.columns if col.endswith('QC')]

df=see.drop(columns=columns_to_drop)

print(var_list)

need_to_drop=['LE_F_MDS','LE_CORR','LE_RANDUNC','H_F_MDS','H_CORR','H_RANDUNC','NEE_VUT_REF_RANDUNC',
             'NEE_VUT_25','NEE_VUT_50','NEE_VUT_75','RECO_NT_VUT_REF','RECO_NT_VUT_25','RECO_NT_VUT_50',
             'RECO_NT_VUT_75','GPP_NT_VUT_25','GPP_NT_VUT_50','GPP_NT_VUT_75','RECO_DT_VUT_REF','RECO_DT_VUT_25',
             'RECO_DT_VUT_50','RECO_DT_VUT_75','GPP_DT_VUT_25','GPP_DT_VUT_50','GPP_DT_VUT_75','RECO_SR',
             'RECO_SR_N','LE_CORR_25','LE_CORR_75','H_CORR_25','H_CORR_75','ID','VARIABLE','PARAMETER',
             'VALUE','TA_ERA_DAY_SD','EBC_CF_N','EBC_CF_METHOD','NIGHT_RANDUNC_N','DAY_RANDUNC_N','NEE_CUT_REF_RANDUNC',
             'NEE_CUT_REF_JOINTUNC','NEE_VUT_REF_JOINTUNC','NEE_CUT_USTAR50_RANDUNC','NEE_VUT_USTAR50_RANDUNC',
             'NEE_CUT_USTAR50_JOINTUNC','NEE_VUT_USTAR50_JOINTUNC','NEE_CUT_MEAN','NEE_VUT_MEAN','NEE_CUT_SE',
             'NEE_VUT_SE','NEE_CUT_05','NEE_CUT_16','NEE_CUT_25','NEE_CUT_50','NEE_CUT_75','NEE_CUT_84','NEE_CUT_95',
             'NEE_CUT_REF_NIGHT_SD', 'NEE_VUT_REF_NIGHT_SD', 'NEE_CUT_REF_NIGHT_RANDUNC', 'NEE_VUT_REF_NIGHT_RANDUNC',
             'NEE_CUT_REF_NIGHT_JOINTUNC', 'NEE_VUT_REF_NIGHT_JOINTUNC','NEE_CUT_REF_DAY_SD', 'NEE_VUT_REF_DAY_SD',
              'NEE_CUT_REF_DAY_RANDUNC', 'NEE_VUT_REF_DAY_RANDUNC', 'NEE_CUT_REF_DAY_JOINTUNC', 'NEE_VUT_REF_DAY_JOINTUNC',
             'NEE_CUT_USTAR50_NIGHT_SD', 'NEE_VUT_USTAR50_NIGHT_SD', 'NEE_CUT_USTAR50_NIGHT_RANDUNC', 'NEE_VUT_USTAR50_NIGHT_RANDUNC',
              'NEE_CUT_USTAR50_NIGHT_JOINTUNC', 'NEE_VUT_USTAR50_NIGHT_JOINTUNC','NEE_CUT_USTAR50_DAY_SD', 'NEE_VUT_USTAR50_DAY_SD', 
              'NEE_CUT_USTAR50_DAY_RANDUNC', 'NEE_VUT_USTAR50_DAY_RANDUNC', 'NEE_CUT_USTAR50_DAY_JOINTUNC', 'NEE_VUT_USTAR50_DAY_JOINTUNC',
             'NEE_CUT_05_NIGHT', 'NEE_CUT_16_NIGHT', 'NEE_CUT_25_NIGHT', 'NEE_CUT_50_NIGHT', 'NEE_CUT_75_NIGHT', 'NEE_CUT_84_NIGHT', 'NEE_CUT_95_NIGHT', 
              'NEE_VUT_05_NIGHT', 'NEE_VUT_16_NIGHT', 'NEE_VUT_25_NIGHT', 'NEE_VUT_50_NIGHT', 'NEE_VUT_75_NIGHT', 'NEE_VUT_84_NIGHT', 'NEE_VUT_95_NIGHT',
              'NEE_CUT_05_DAY', 'NEE_CUT_16_DAY', 'NEE_CUT_25_DAY', 'NEE_CUT_50_DAY', 'NEE_CUT_75_DAY', 'NEE_CUT_84_DAY', 'NEE_CUT_95_DAY', 'NEE_VUT_05_DAY',
              'NEE_VUT_16_DAY', 'NEE_VUT_25_DAY', 'NEE_VUT_50_DAY', 'NEE_VUT_75_DAY', 'NEE_VUT_84_DAY', 'NEE_VUT_95_DAY', 'RECO_NT_VUT_USTAR50', 'RECO_NT_VUT_MEAN', 
              'RECO_NT_VUT_SE', 'RECO_NT_VUT_05', 'RECO_NT_VUT_16', 'RECO_NT_VUT_84', 'RECO_NT_VUT_95', 'RECO_NT_CUT_REF', 'RECO_NT_CUT_USTAR50', 'RECO_NT_CUT_MEAN', 
              'RECO_NT_CUT_SE', 'RECO_NT_CUT_05', 'RECO_NT_CUT_16', 'RECO_NT_CUT_25', 'RECO_NT_CUT_50', 'RECO_NT_CUT_75', 'RECO_NT_CUT_84', 'RECO_NT_CUT_95', 
             'GPP_NT_VUT_SE', 'GPP_NT_VUT_05', 'GPP_NT_VUT_16', 'GPP_NT_VUT_84', 'GPP_NT_VUT_95', 'GPP_NT_CUT_SE', 'GPP_NT_CUT_05', 'GPP_NT_CUT_16', 'GPP_NT_CUT_25', 
              'GPP_NT_CUT_50', 'GPP_NT_CUT_75', 'GPP_NT_CUT_84', 'GPP_NT_CUT_95', 'RECO_DT_VUT_USTAR50', 'RECO_DT_VUT_MEAN', 'RECO_DT_VUT_SE', 'RECO_DT_VUT_05', 'RECO_DT_VUT_16', 
              'RECO_DT_VUT_84', 'RECO_DT_VUT_95', 'RECO_DT_CUT_REF', 'RECO_DT_CUT_USTAR50', 'RECO_DT_CUT_MEAN', 'RECO_DT_CUT_SE', 'RECO_DT_CUT_05', 'RECO_DT_CUT_16', 'RECO_DT_CUT_25', 
              'RECO_DT_CUT_50', 'RECO_DT_CUT_75', 'RECO_DT_CUT_84', 'RECO_DT_CUT_95','GPP_DT_VUT_SE', 'GPP_DT_VUT_05', 'GPP_DT_VUT_16', 'GPP_DT_VUT_84', 'GPP_DT_VUT_95', 'GPP_DT_CUT_REF',
             'GPP_DT_CUT_SE', 'GPP_DT_CUT_05', 'GPP_DT_CUT_16', 'GPP_DT_CUT_25', 'GPP_DT_CUT_50', 'GPP_DT_CUT_75', 'GPP_DT_CUT_84', 'GPP_DT_CUT_95',
             'LE_CORR_JOINTUNC', 'H_CORR_JOINTUNC', 'NIGHT_D', 'DAY_D', 'TS_F_MDS_5', 'TS_F_MDS_6', 'SWC_F_MDS_4', 'TS_F_MDS_7', 'SWC_F_MDS_5', 'SWC_F_MDS_6', 'SWC_F_MDS_7', 'TS_F_MDS_8', 'TS_F_MDS_9']

df=df.drop(columns=need_to_drop)

df.head()

df.to_csv('/Users/abigailbase/PROJECT FILES/FINAL DF/final_df.csv')
