==> 3.ERA5_Land_data_retrieval .ipynb <==
# This file contains the code that was ran to retrieve the ERA5-Land data from the Climate Data Store (CDS) API

# Link to the data: https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-land?tab=overview

import cdsapi

c=cdsapi.Client()

#AR-Vir

latitude = -28.2395
longitude = -56.1886
bounding_box = [
    latitude + 0.01, longitude - 0.01,  
    latitude - 0.01, longitude + 0.01   
]

#2009 

c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            'surface_solar_radiation_downwards'
        ],
        'year': ['2010','2011','2012'],  
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'],  
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ],  
        'time': [
            '12:00'
        ],  
        'area': bounding_box,  
        'format': 'netcdf'
    },
    'era5_AR-Vir_1.nc' 
)

#2010

c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            '2m_temperature', '2m_dewpoint_temperature', 'surface_pressure', 'surface_solar_radiation_downwards',
            'total_precipitation', '10m_u_component_of_wind', '10m_v_component_of_wind',
            'temperature_of_soil_layer_1', 'volumetric_soil_water_layer_1'
        ],
        'year': '2010',  
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'],
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ], 
        'time': [
            '12:00'
        ],  
        'area': bounding_box,  
        'format': 'netcdf' 
    },
    'era5_AR-Vir_2.nc' 
)

#2011

c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            '2m_temperature', '2m_dewpoint_temperature', 'surface_pressure', 'surface_solar_radiation_downwards',
            'total_precipitation', '10m_u_component_of_wind', '10m_v_component_of_wind',
            'temperature_of_soil_layer_1', 'volumetric_soil_water_layer_1'
        ],
        'year': '2011', 
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'],  
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ], 
        'time': [
            '12:00'
        ],  
        'area': bounding_box,  
        'format': 'netcdf' 
    },
    'era5_AR-Vir_3.nc' 
)

#2012

c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            '2m_temperature', '2m_dewpoint_temperature', 'surface_pressure', 'surface_solar_radiation_downwards',
            'total_precipitation', '10m_u_component_of_wind', '10m_v_component_of_wind',
            'temperature_of_soil_layer_1', 'volumetric_soil_water_layer_1'
        ],
        'year': '2012',  
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'], 
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ], 
        'time': [
            '12:00'
        ],  
        'area': bounding_box,  
        'format': 'netcdf' 
    },
    'era5_AR-Vir_4.nc'  
)

#2013

c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            '2m_temperature', '2m_dewpoint_temperature', 'surface_pressure', 'surface_solar_radiation_downwards',
            'total_precipitation', '10m_u_component_of_wind', '10m_v_component_of_wind',
            'temperature_of_soil_layer_1', 'volumetric_soil_water_layer_1'
        ],
        'year': '2013', 
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'],  
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ], 
        'time': [
            '12:00'
        ], 
        'area': bounding_box,  
        'format': 'netcdf'  
    },
    'era5_AR-Vir_5.nc' 
)

#2014

c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            '2m_temperature', '2m_dewpoint_temperature', 'surface_pressure', 'surface_solar_radiation_downwards',
            'total_precipitation', '10m_u_component_of_wind', '10m_v_component_of_wind',
            'temperature_of_soil_layer_1', 'volumetric_soil_water_layer_1'
        ],
        'year': '2014',  
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'], 
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ],  
        'time': [
            '12:00'
        ],  
        'area': bounding_box,  
        'format': 'netcdf'  
    },
    'era5_AR-Vir_6.nc'  
)

### AU-Dry

latitude = -15.2588
longitude = 132.3706
bounding_box = [
    latitude + 0.01, longitude - 0.01,  
    latitude - 0.01, longitude + 0.01   
]

c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            'surface_solar_radiation_downwards'],
        'year': ['2010','2011','2012','2013','2014'],  
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'],  # All months
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ], 
        'time': [
            '12:00'
        ],  
        'area': bounding_box,  
        'format': 'netcdf'  
    },
    'era5_AU-Dry_1.nc'  
)

###BE-Vie 

latitude = 50.3049
longitude = 5.9981
bounding_box = [
    latitude + 0.01, longitude - 0.01,  
    latitude - 0.01, longitude + 0.01  
]

c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            'surface_solar_radiation_downwards'],
        'year': ['2010','2011','2012','2013','2014'], 
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'],  
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ], 
        'time': [
            '12:00'
        ],  
        'area': bounding_box,  
        'format': 'netcdf'  
    },
    'era5_BE-Vie_1.nc'  
)

### CA-TP1

latitude = 42.6609
longitude = -80.5595
bounding_box = [
    latitude + 0.01, longitude - 0.01,  
    latitude - 0.01, longitude + 0.01   
]

c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            'surface_solar_radiation_downwards'],
        'year': ['2010','2011','2012','2013','2014'],  
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'], 
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ], 
        'time': [
            '12:00'
        ], 
        'area': bounding_box,
        'format': 'netcdf'  
    },
    'era5_CA-TP1_1.nc'  
)

### CH-Cha

latitude = 47.2102
longitude = 8.4104
bounding_box = [
    latitude + 0.01, longitude - 0.01,  
    latitude - 0.01, longitude + 0.01   
]

c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            'surface_solar_radiation_downwards'],
        'year': ['2010','2011','2012','2013','2014'],  
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'],  
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ],  
        'time': [
            '12:00'
        ],  
        'area': bounding_box,  
        'format': 'netcdf' 
    },
    'era5_CH-Cha_1.nc'  
)

### DE-Gri

latitude = 50.9500
longitude = 13.5126
bounding_box = [
    latitude + 0.01, longitude - 0.01, 
    latitude - 0.01, longitude + 0.01   
]

c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            'surface_solar_radiation_downwards'],
        'year': ['2010','2011','2012','2013','2014'],  
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'],  
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ],  
        'time': [
            '12:00'
        ], 
        'area': bounding_box,  
        'format': 'netcdf'  
    },
    'era5_DE-Gri_1.nc' 
)

### FR-Pue


latitude = 43.7413
longitude = 3.5957
bounding_box = [
    latitude + 0.01, longitude - 0.01,  
    latitude - 0.01, longitude + 0.01   
]



c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            'surface_solar_radiation_downwards'],
        'year': ['2010','2011','2012','2013','2014'], 
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'],  
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ],  
        'time': [
            '12:00'
        ],  
        'area': bounding_box, 
        'format': 'netcdf'  
    },
    'era5_FR-Pue_1.nc'  
)

### GF-Guy

latitude = 5.2788
longitude = -52.9249
bounding_box = [
    latitude + 0.01, longitude - 0.01,  
    latitude - 0.01, longitude + 0.01  
]


c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            'surface_solar_radiation_downwards'],
        'year': ['2010','2011','2012','2013','2014'], 
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'],  
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ], 
        'time': [
            '12:00'
        ], 
        'area': bounding_box, 
        'format': 'netcdf'  
    },
    'era5_GF-Guy_1.nc' 
)

### IT-Col

latitude = 41.8494
longitude = 13.5881
bounding_box = [
    latitude + 0.01, longitude - 0.01,  
    latitude - 0.01, longitude + 0.01  
]

c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            'surface_solar_radiation_downwards'],
        'year': ['2010','2011','2012','2013','2014'], 
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'],  
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ], 
        'time': [
            '12:00'
        ],  
        'area': bounding_box,  
        'format': 'netcdf'  
    },
    'era5_IT-Col_1.nc'  
)

### NL-Loo

latitude = 52.1666
longitude = 5.7436
bounding_box = [
    latitude + 0.01, longitude - 0.01,  
    latitude - 0.01, longitude + 0.01   
]

c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            'surface_solar_radiation_downwards'],
        'year': ['2010','2011','2012','2013','2014'],  
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'], 
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ],  
        'time': [
            '12:00'
        ], 
        'area': bounding_box, 
        'format': 'netcdf'  
    },
    'era5_NL-Loo_1.nc' 
)

### RU-Cok

latitude = 70.8291
longitude = 147.4943
bounding_box = [
    latitude + 0.01, longitude - 0.01,  
    latitude - 0.01, longitude + 0.01   
]

c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            'surface_solar_radiation_downwards'],
        'year': ['2010','2011','2012','2013','2014'],  
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'], 
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ],  
        'time': [
            '12:00'
        ],  
        'area': bounding_box, 
        'format': 'netcdf'  
    },
    'era5_RU-Cok_1.nc' 
)

### RU-Fyo

latitude = 56.4615
longitude = 32.9221
bounding_box = [
    latitude + 0.01, longitude - 0.01, 
    latitude - 0.01, longitude + 0.01  
]

c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            'surface_solar_radiation_downwards'],
        'year': ['2010','2011','2012','2013','2014'],  
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'],  
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ],  
        'time': [
            '12:00'
        ], 
        'area': bounding_box,
        'format': 'netcdf'  
    },
    'era5_RU-Fyo_1.nc' 
)

### US-PFa

latitude = 45.9459
longitude = -90.2723
bounding_box = [
    latitude + 0.01, longitude - 0.01, 
    latitude - 0.01, longitude + 0.01   
]

c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            'surface_solar_radiation_downwards'],
        'year': ['2010','2011','2012','2013','2014'],  
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'],  
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ], 
        'time': [
            '12:00'
        ], 
        'area': bounding_box,  
        'format': 'netcdf'  
    },
    'era5_US-PFa_1.nc' 
)

### US-Var

latitude = 38.4133
longitude = -120.9508
bounding_box = [
    latitude + 0.01, longitude - 0.01,  
    latitude - 0.01, longitude + 0.01   
]

c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            'surface_solar_radiation_downwards'],
        'year': ['2010','2011','2012','2013','2014'],  
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'], 
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ],  
        'time': [
            '12:00'
        ],  
        'area': bounding_box,  
        'format': 'netcdf'  
    },
    'era5_US-Var_1.nc'  
)

### ZA-Kru

latitude = -25.0197
longitude = 31.4969
bounding_box = [
    latitude + 0.01, longitude - 0.01,  
    latitude - 0.01, longitude + 0.01   
]

c.retrieve(
    'reanalysis-era5-land',
    {
        'variable': [
            'surface_solar_radiation_downwards'],
        'year': ['2010','2011','2012','2013','2014'], 
        'month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'],  
        'day': [
            '01', '02', '03', '04', '05', '06', '07',
            '08', '09', '10', '11', '12', '13', '14',
            '15', '16', '17', '18', '19', '20', '21',
            '22', '23', '24', '25', '26', '27', '28',
            '29', '30', '31'
        ],  
        'time': [
            '12:00'
        ],  
        'area': bounding_box,  
        'format': 'netcdf'  
    },
    'era5_ZA-Kru_1.nc' 
)
